Update counter ABI and related configurations
- Renamed `multi_ibs_counter` to `bls_multisig_counter` in counter.abi.ts and updated all related references.
- Updated struct names from `MultiIBSConfig` to `BlsMultisigConfig` and `MultiIBSCounter` to `BlsMultisigCounter`.
- Added new field `public_key_bytes` to `BlsAggregatedPk` struct.
- Updated generated timestamps in ABI files.
- Updated package ID in suigen-config.json to reflect new address for the counter package.
