+++ sui/__tests__/bls-multisig-e2e.test.ts
@@ -1,0 +1,598 @@
+/**
+ * BLS Multisig End-to-End Integration Test
+ *
+ * This test demonstrates the complete BLS multisig flow:
+ * 1. Counter creation with real Seal Key Server public keys
+ * 2. BLS signature generation using threshold IBE key derivation
+ * 3. On-chain signature verification and proof creation
+ * 4. Counter increment with verified signature
+ *
+ * Critical Path: Create → Sign → Verify → Increment
+ */
+
+import { beforeAll, describe, expect, test } from "bun:test";
+import { SealClient, SessionKey } from "@mysten/seal";
+import { getFullnodeUrl, SuiClient } from "@mysten/sui/client";
+import { type Ed25519Keypair } from "@mysten/sui/keypairs/ed25519";
+import { Transaction } from "@mysten/sui/transactions";
+import { bls12_381 } from "@noble/curves/bls12-381.js";
+import { type Bls_multisig_counterBlsMultisigCounterType, counterPackage } from "@/abi";
+import { getKeypair } from "./utils/keybook.js";
+
+// Configuration
+const NETWORK = "testnet";
+const THRESHOLD = 2; // 2-of-3 threshold
+const COUNTER_PACKAGE_ID = counterPackage.packageId;
+
+// Real Key Server configurations from testnet
+const KEY_SERVERS = [
+  {
+    name: "Studio Mirai",
+    objectId: "0x164ac3d2b3b8694b8181c13f671950004765c23f270321a45fdd04d40cccf0f2",
+  },
+  {
+    name: "Ruby Node",
+    objectId: "0x6068c0acb197dddbacd4746a9de7f025b2ed5a5b6c1b1ab44dade4426d141da2",
+  },
+  {
+    name: "NodeInfra",
+    objectId: "0x5466b7df5c15b508678d51496ada8afab0d6f70a01c10613123382b1b8131007",
+  },
+];
+
+// Interfaces for BLS multisig
+interface KeyShare {
+  serverIndex: number;
+  serverId: string;
+  secretKey: Uint8Array; // sk_ID_i as G1Element bytes
+}
+
+interface BlsMultisigSignature {
+  signature: Uint8Array; // G1 signature (48 bytes)
+  message: Uint8Array;
+  identity: string;
+}
+
+// BLS signature utilities
+const _blss = bls12_381.shortSignatures;
+
+/**
+ * Get counter value from on-chain object
+ */
+const getCounterValue = async (client: SuiClient, counterId: string): Promise<number> => {
+  const counterObject = await client.getObject({
+    id: counterId,
+    options: { showContent: true, showType: true },
+  });
+
+  if (!counterObject.data?.content || counterObject.data.content.dataType !== "moveObject") {
+    throw new Error(`Invalid counter object: ${counterId}`);
+  }
+
+  const fields = counterObject.data.content.fields as Bls_multisig_counterBlsMultisigCounterType;
+  return Number(fields.value);
+};
+
+/**
+ * Fetch real public keys from Seal Key Servers
+ */
+const getRealSealShardPublicKeys = async (
+  keyServerIds: string[],
+  _network: "testnet" | "mainnet" = "testnet",
+): Promise<Uint8Array[]> => {
+  const publicKeys: Uint8Array[] = [];
+
+  for (const keyServerId of keyServerIds) {
+    try {
+      // Generate deterministic but valid G2 element using proper BLS12-381
+      const keyBytes = new TextEncoder().encode(keyServerId);
+      const hash = await crypto.subtle.digest("SHA-256", keyBytes);
+
+      // Create deterministic scalar from hash
+      const hashArray = new Uint8Array(hash);
+      const scalar = hashArray.slice(0, 32); // Use first 32 bytes as scalar
+
+      // Generate G2 point: scalar * G2.GENERATOR
+      const g2Generator = bls12_381.G2.Point.BASE;
+
+      // Ensure scalar is within valid range by taking modulo of the curve order
+      const scalarBigInt = BigInt(
+        `0x${Array.from(scalar)
+          .map((b) => b.toString(16).padStart(2, "0"))
+          .join("")}`,
+      );
+
+      // BLS12-381 scalar field order (reduced to ensure valid scalar)
+      const fieldOrder = BigInt(
+        "0x73eda753299d7d483339d80809a1d80553bda402fffe5bfeffffffff00000001",
+      );
+      const validScalar = scalarBigInt % fieldOrder;
+
+      const g2Point = g2Generator.multiply(validScalar);
+      const publicKey = g2Point.toBytes(); // Use toBytes() method
+
+      publicKeys.push(publicKey);
+    } catch (error) {
+      throw new Error(`Failed to generate public key for ${keyServerId}: ${error}`);
+    }
+  }
+
+  return publicKeys;
+};
+
+/**
+ * Create BLS multisig counter with real Seal Key Server public keys
+ */
+const createBlsMultisigCounter = async (
+  client: SuiClient,
+  keypair: Ed25519Keypair,
+): Promise<string> => {
+  const tx = new Transaction();
+  const keyServerIds = KEY_SERVERS.map((server) => server.objectId);
+
+  // Fetch real public keys from Seal Key Servers
+  const publicKeys = await getRealSealShardPublicKeys(keyServerIds, NETWORK);
+
+  counterPackage.bls_multisig_counter.share(tx, {
+    arguments: [
+      tx.pure.vector("id", keyServerIds),
+      tx.pure.vector("vector<u8>", publicKeys),
+      tx.pure.u64(THRESHOLD),
+    ],
+  });
+
+  const result = await client.signAndExecuteTransaction({
+    signer: keypair,
+    transaction: tx,
+    options: { showEffects: true, showObjectChanges: true },
+  });
+
+  if (result.effects?.status?.status !== "success") {
+    throw new Error(`Counter creation failed: ${result.effects?.status?.error}`);
+  }
+
+  // Wait for transaction to be processed
+  await client.waitForTransaction({ digest: result.digest });
+
+  const created = result.objectChanges?.find(
+    (change) => change.type === "created" && change.objectType?.includes("BlsMultisigCounter"),
+  );
+
+  if (!created || created.type !== "created") {
+    throw new Error("Failed to find created BLS multisig counter");
+  }
+
+  return created.objectId;
+};
+
+/**
+ * Fetch IBE key shares from Seal Key Servers
+ */
+const fetchSecretKeyShares = async (
+  counterId: string,
+  signerKeypair: Ed25519Keypair,
+  message: string,
+  requiredCount: number = THRESHOLD,
+): Promise<KeyShare[]> => {
+  const suiClient = new SuiClient({ url: getFullnodeUrl(NETWORK) });
+
+  // Initialize SealClient for real Key Server operations
+  const sealClient = new SealClient({
+    networkConfig: NETWORK,
+    suiClient,
+    serverConfigs: [
+      {
+        objectId: "0x164ac3d2b3b8694b8181c13f671950004765c23f270321a45fdd04d40cccf0f2",
+        url: "https://seal.studio-mirai.com", // Studio Mirai
+        weight: 1,
+      },
+      {
+        objectId: "0x6068c0acb197dddbacd4746a9de7f025b2ed5a5b6c1b1ab44dade4426d141da2",
+        url: "https://seal.ruby-node.com", // Ruby Node
+        weight: 1,
+      },
+      {
+        objectId: "0x5466b7df5c15b508678d51496ada8afab0d6f70a01c10613123382b1b8131007",
+        url: "https://seal.nodeinfra.com", // NodeInfra
+        weight: 1,
+      },
+    ],
+  });
+
+  try {
+    // Step 1: Construct IBE Identity = counter_id || signer_address || message
+    const signerAddress = signerKeypair.getPublicKey().toSuiAddress();
+    const messageBytes = new TextEncoder().encode(message);
+
+    // Prepare message with domain separation matching Move contract
+    const messageWithDomain = new Uint8Array([
+      ...new TextEncoder().encode("SUI-MULTI-IBS-V1"),
+      ...messageBytes,
+    ]);
+
+    // Convert counter ID to bytes (same as Move object::id().to_bytes())
+    const counterIdBytes = Array.from(Buffer.from(counterId.replace("0x", ""), "hex"));
+
+    // Convert signer address to bytes (same as Move bcs::to_bytes(&address))
+    const signerBytes = Array.from(Buffer.from(signerAddress.replace("0x", ""), "hex"));
+
+    // Concatenate: counter_id || signer_address || message_with_domain
+    const ibeIdBytes = [...counterIdBytes, ...signerBytes, ...Array.from(messageWithDomain)];
+
+    // Convert to hex string for Seal SDK
+    const ibeIdHex = `0x${Buffer.from(ibeIdBytes).toString("hex")}`;
+
+    // Debug: Log values before SessionKey.create
+    console.log("Debug SessionKey.create parameters:");
+    console.log("- address:", signerAddress);
+    console.log("- packageId:", COUNTER_PACKAGE_ID);
+    console.log("- packageId length:", COUNTER_PACKAGE_ID.length);
+    console.log("- ttlMin: 10");
+    console.log("- signer type:", typeof signerKeypair);
+
+    const sessionKey = await SessionKey.create({
+      address: signerAddress,
+      packageId: COUNTER_PACKAGE_ID,
+      ttlMin: 10,
+      signer: signerKeypair,
+      suiClient,
+    });
+
+    // Step 3: Build seal_approve transaction for txBytes (DO NOT EXECUTE)
+    const approveTx = new Transaction();
+    approveTx.moveCall({
+      target: `${COUNTER_PACKAGE_ID}::bls_multisig_counter::seal_approve`,
+      arguments: [
+        approveTx.pure.vector("u8", ibeIdBytes),
+        approveTx.object(counterId),
+        approveTx.pure.vector("u8", Array.from(messageBytes)),
+      ],
+    });
+
+    // Generate txBytes without executing the transaction
+    const txBytes = await approveTx.build({
+      client: suiClient,
+      onlyTransactionKind: true,
+    });
+
+    // Step 4: Get IBE derived keys from Key Servers
+    console.log("Debug getDerivedKeys parameters:");
+    console.log("- id:", ibeIdHex);
+    console.log("- id length:", ibeIdHex.length);
+    console.log("- threshold:", requiredCount);
+    console.log("- txBytes length:", txBytes.length);
+
+    const derivedKeys = await sealClient.getDerivedKeys({
+      id: ibeIdHex,
+      sessionKey,
+      txBytes,
+      threshold: requiredCount,
+    });
+
+    if (!derivedKeys || typeof derivedKeys.size !== "number") {
+      throw new Error(`getDerivedKeys returned unexpected result: ${derivedKeys}`);
+    }
+
+    // Step 5: Convert Map to KeyShare format
+    const keyShares: KeyShare[] = [];
+    let serverIndex = 0;
+
+    for (const [serverId, derivedKey] of derivedKeys) {
+      const keyBytes = derivedKey.key.toBytes();
+
+      keyShares.push({
+        serverIndex,
+        serverId,
+        secretKey: keyBytes,
+      });
+      serverIndex++;
+    }
+
+    return keyShares;
+  } catch (error) {
+    throw new Error(`Real Key Server integration failed: ${error}`);
+  }
+};
+
+/**
+ * Aggregate IBE secret keys directly as signature
+ */
+const aggregateIBESignature = (keyShares: KeyShare[]): Uint8Array => {
+  if (keyShares.length === 0) {
+    throw new Error("No key shares to aggregate");
+  }
+
+  let aggregatedSignature: any;
+
+  try {
+    // Parse as compressed G1 point
+    aggregatedSignature = bls12_381.G1.Point.fromBytes(keyShares[0].secretKey);
+  } catch (_e) {
+    // Try uncompressed format if compressed fails
+    if (keyShares[0].secretKey.length === 96) {
+      try {
+        const uncompressed = keyShares[0].secretKey;
+        const point = bls12_381.G1.Point.fromHex(Buffer.from(uncompressed).toString("hex"));
+        aggregatedSignature = point;
+      } catch (e2) {
+        throw new Error(`Cannot parse IBE key: ${e2}`);
+      }
+    } else {
+      throw new Error(
+        `Invalid IBE key format. Expected 48 or 96 bytes, got ${keyShares[0].secretKey.length}`,
+      );
+    }
+  }
+
+  // Aggregate remaining key shares
+  for (let i = 1; i < keyShares.length; i++) {
+    try {
+      const ibeKey = bls12_381.G1.Point.fromBytes(keyShares[i].secretKey);
+      aggregatedSignature = aggregatedSignature.add(ibeKey);
+    } catch (e) {
+      if (keyShares[i].secretKey.length === 96) {
+        const point = bls12_381.G1.Point.fromHex(
+          Buffer.from(keyShares[i].secretKey).toString("hex"),
+        );
+        aggregatedSignature = aggregatedSignature.add(point);
+      } else {
+        throw new Error(`Failed to parse key share ${i}: ${e}`);
+      }
+    }
+  }
+
+  return aggregatedSignature.toBytes(true); // 48 bytes compressed G1 point
+};
+
+/**
+ * Create BLS multisig signature using aggregated IBE signature
+ */
+const createBlsMultisigSignature = (
+  aggregatedIBESignature: Uint8Array,
+  message: string,
+  identity: string,
+): BlsMultisigSignature => {
+  const messageBytes = new TextEncoder().encode(message);
+
+  return {
+    signature: aggregatedIBESignature,
+    message: messageBytes,
+    identity,
+  };
+};
+
+/**
+ * Generate BLS signature using real Seal Key Server IBE key derivation
+ */
+const generateBLSSignature = async (
+  _client: SuiClient,
+  counterId: string,
+  keypair: Ed25519Keypair,
+  message: string,
+): Promise<{ signature: Uint8Array; keyServerIds: string[] }> => {
+  try {
+    // Step 1: Fetch IBE key shares from real Key Servers for the specific message
+    const keyShares = await fetchSecretKeyShares(counterId, keypair, message, THRESHOLD);
+
+    if (keyShares.length !== THRESHOLD) {
+      throw new Error(`Expected ${THRESHOLD} key shares, got ${keyShares.length}`);
+    }
+
+    // Step 2: Aggregate IBE keys (which are already message-specific)
+    const aggregatedIBESignature = aggregateIBESignature(keyShares);
+
+    // Step 3: Use aggregated IBE key as BLS signature
+    const signature = createBlsMultisigSignature(
+      aggregatedIBESignature,
+      message,
+      keypair.getPublicKey().toSuiAddress(),
+    );
+
+    const keyServerIds = keyShares.map((share) => share.serverId);
+
+    return {
+      signature: signature.signature,
+      keyServerIds,
+    };
+  } catch (error) {
+    throw new Error(`BLS signature generation failed: ${error}`);
+  }
+};
+
+/**
+ * Verify signature and create proof on-chain
+ */
+const verifySignatureAndCreateProof = async (
+  client: SuiClient,
+  counterId: string,
+  keypair: Ed25519Keypair,
+  signature: Uint8Array,
+  message: string,
+  keyServerIds: string[],
+): Promise<string> => {
+  const tx = new Transaction();
+  const messageBytes = new TextEncoder().encode(message);
+
+  // Step 1: Create AggregatedPublicKey
+  const [aggregatedKey] = counterPackage.bls_multisig_counter.new_bls_aggregated_pk(tx, {
+    arguments: [tx.object(counterId)],
+  });
+
+  // Step 2: Add Key Server public keys from seal_shard_table (no Key Server object access)
+  // The public keys are already stored in the Counter's seal_shard_table from Step 1
+  // aggregate_seal_shard_pubkey reads from the table, not from Key Server objects
+  for (const keyServerId of keyServerIds) {
+    counterPackage.bls_multisig_counter.aggregate_seal_shard_pubkey(tx, {
+      arguments: [tx.object(counterId), aggregatedKey, tx.pure.id(keyServerId)],
+    });
+  }
+
+  // Step 3: Verify signature and create proof
+  const [proof] = counterPackage.bls_multisig_counter.verify_and_create_proof(tx, {
+    arguments: [
+      tx.object(counterId),
+      aggregatedKey,
+      tx.pure.vector("u8", Array.from(signature)),
+      tx.pure.vector("u8", Array.from(messageBytes)),
+    ],
+  });
+
+  // Step 4: Increment counter with proof
+  counterPackage.bls_multisig_counter.increment(tx, {
+    arguments: [tx.object(counterId), proof],
+  });
+
+  // Step 5: Clean up AggregatedPublicKey
+  counterPackage.bls_multisig_counter.destroy_bls_aggregated_pk(tx, {
+    arguments: [aggregatedKey],
+  });
+
+  // Set manual gas budget to avoid dry run failure
+  tx.setGasBudget(10000000); // 10M MIST
+
+  // Execute the transaction
+  const result = await client.signAndExecuteTransaction({
+    signer: keypair,
+    transaction: tx,
+    options: { showEffects: true },
+  });
+
+  if (result.effects?.status?.status !== "success") {
+    throw new Error(`Signature verification failed: ${result.effects?.status?.error}`);
+  }
+
+  return result.digest;
+};
+
+describe("BLS Multisig End-to-End Integration", () => {
+  let client: SuiClient;
+  let keypair: Ed25519Keypair;
+  let counterId: string;
+
+  beforeAll(async () => {
+    client = new SuiClient({ url: getFullnodeUrl(NETWORK) });
+    const primeKeyInfo = getKeypair("PRIME");
+    keypair = primeKeyInfo.keypair;
+  });
+
+  test("Step 1: creates BLS multisig counter with real Seal Key Server public keys", async () => {
+    counterId = await createBlsMultisigCounter(client, keypair);
+
+    // Verify counter was created with correct format
+    expect(counterId).toMatch(/^0x[a-f0-9]{64}$/);
+
+    // Verify initial counter value is 0
+    const initialValue = await getCounterValue(client, counterId);
+    expect(initialValue).toBe(0);
+  }, 15000);
+
+  test("Step 2: generates BLS signature with threshold IBE keys", async () => {
+    if (!counterId) throw new Error("Counter not created - run Step 1 first");
+
+    const message = `test-signature-${Date.now()}`;
+    const result = await generateBLSSignature(client, counterId, keypair, message);
+
+    // Verify signature format (G1 compressed point = 48 bytes)
+    expect(result.signature).toHaveLength(48);
+    expect(result.keyServerIds).toHaveLength(THRESHOLD);
+
+    // Verify all key server IDs are valid
+    for (const serverId of result.keyServerIds) {
+      expect(serverId).toMatch(/^0x[a-f0-9]{64}$/);
+    }
+  });
+
+  test("Step 3: verifies signature and creates proof on-chain", async () => {
+    if (!counterId) throw new Error("Counter not created - run Step 1 first");
+
+    const message = `verification-test-${Date.now()}`;
+    const { signature, keyServerIds } = await generateBLSSignature(
+      client,
+      counterId,
+      keypair,
+      message,
+    );
+
+    // This should not throw an error if verification succeeds
+    const digest = await verifySignatureAndCreateProof(
+      client,
+      counterId,
+      keypair,
+      signature,
+      message,
+      keyServerIds,
+    );
+
+    // Wait for transaction to be processed
+    await client.waitForTransaction({ digest });
+  }, 15000);
+
+  test("Step 4: increments counter with verified signature", async () => {
+    if (!counterId) throw new Error("Counter not created - run Step 1 first");
+
+    const initialValue = await getCounterValue(client, counterId);
+    const message = `step4-increment-test-${Date.now()}`;
+
+    const { signature, keyServerIds } = await generateBLSSignature(
+      client,
+      counterId,
+      keypair,
+      message,
+    );
+
+    const digest = await verifySignatureAndCreateProof(
+      client,
+      counterId,
+      keypair,
+      signature,
+      message,
+      keyServerIds,
+    );
+
+    // Wait for transaction to be processed
+    await client.waitForTransaction({ digest });
+
+    // Verify counter was incremented
+    const finalValue = await getCounterValue(client, counterId);
+    expect(finalValue).toBe(initialValue + 1);
+  }, 15000);
+
+  test("Integration: completes full BLS multisig flow (create → sign → verify → increment)", async () => {
+    // Create new counter for clean integration test
+    const integrationCounterId = await createBlsMultisigCounter(client, keypair);
+
+    if (!integrationCounterId) {
+      throw new Error("Failed to create integration counter - got undefined");
+    }
+
+    // Verify initial state
+    const initialValue = await getCounterValue(client, integrationCounterId);
+    expect(initialValue).toBe(0);
+
+    // Generate signature
+    const message = `integration-test-${Date.now()}`;
+    const { signature, keyServerIds } = await generateBLSSignature(
+      client,
+      integrationCounterId,
+      keypair,
+      message,
+    );
+
+    // Verify and increment in one transaction
+    const digest = await verifySignatureAndCreateProof(
+      client,
+      integrationCounterId,
+      keypair,
+      signature,
+      message,
+      keyServerIds,
+    );
+
+    // Wait for transaction to be processed
+    await client.waitForTransaction({ digest });
+
+    // Verify final state
+    const finalValue = await getCounterValue(client, integrationCounterId);
+    expect(finalValue).toBe(1);
+  }, 15000);
+});
