--- sui/__tests__/utils/real-seal-keys.ts
@@ -1,154 +1,0 @@
-/**
- * Real Seal Key Server Integration Utilities
- *
- * This module provides utilities to fetch real public keys from Seal Key Servers
- * and completely eliminates mock usage in tests.
- */
-
-import { SealMultiIBSAggregator } from "../../scripts/multi-ibs-with-seal.ts";
-
-// Cache for real public keys to avoid repeated network calls
-const publicKeyCache = new Map<string, Uint8Array>();
-
-/**
- * Fetch real public keys from Seal Key Servers
- * Uses caching to improve test performance
- * Falls back to deterministic key generation if network fetch fails
- */
-export async function getRealSealShardPublicKeys(
-  keyServerIds: string[],
-  network: "testnet" | "mainnet" = "testnet",
-): Promise<Uint8Array[]> {
-  const publicKeys: Uint8Array[] = [];
-
-  for (const keyServerId of keyServerIds) {
-    // Check cache first
-    let publicKey = publicKeyCache.get(keyServerId);
-
-    if (!publicKey) {
-      try {
-        // Try to fetch from real Seal Key Server
-        const aggregator = new SealMultiIBSAggregator(network);
-        publicKey = await aggregator.getSealShardPublicKey(keyServerId);
-        // Cache the successful result
-        publicKeyCache.set(keyServerId, publicKey);
-      } catch (_error) {
-        // Generate deterministic but valid G2 element using proper BLS12-381
-        const { bls12_381 } = await import("@noble/curves/bls12-381.js");
-        const keyBytes = new TextEncoder().encode(keyServerId);
-        const hash = await crypto.subtle.digest("SHA-256", keyBytes);
-
-        // Create deterministic scalar from hash
-        const hashArray = new Uint8Array(hash);
-        const scalar = hashArray.slice(0, 32); // Use first 32 bytes as scalar
-
-        // Generate G2 point: scalar * G2.GENERATOR
-        const g2Generator = bls12_381.G2.Point.BASE;
-        const scalarBigInt = BigInt(
-          `0x${Array.from(scalar)
-            .map((b) => b.toString(16).padStart(2, "0"))
-            .join("")}`,
-        );
-        // Reduce modulo curve order to ensure valid scalar
-        const reducedScalar = scalarBigInt % bls12_381.fields.Fr.ORDER;
-        const scalaredPoint = g2Generator.multiply(reducedScalar);
-
-        // Convert to compressed bytes (96 bytes for G2)
-        publicKey = scalaredPoint.toBytes(true);
-
-        // Cache the deterministic result
-        publicKeyCache.set(keyServerId, publicKey);
-      }
-    }
-
-    publicKeys.push(publicKey);
-  }
-
-  return publicKeys;
-}
-
-/**
- * Create real seal shard counter using actual public keys from Seal Key Servers
- */
-export async function createRealSealShardCounter(
-  keyServerIds: string[],
-  _threshold: number,
-  network: "testnet" | "mainnet" = "testnet",
-): Promise<{ keyServerIds: string[]; publicKeys: number[][] }> {
-  // Fetch real public keys from Seal Key Servers
-  const realPublicKeys = await getRealSealShardPublicKeys(keyServerIds, network);
-
-  return {
-    keyServerIds,
-    publicKeys: realPublicKeys.map((pk) => Array.from(pk)),
-  };
-}
-
-/**
- * Generate real BLS signature using actual cryptographic operations
- * NO MOCKS - uses real cryptographic libraries
- */
-export async function createRealBLSSignature(
-  message: string,
-  keyServerIds: string[],
-  threshold: number,
-  _network: "testnet" | "mainnet" = "testnet",
-): Promise<{ signature_g1: number[]; message: number[] }> {
-  const messageBytes = new TextEncoder().encode(message);
-
-  try {
-    // Use actual cryptographic libraries for real signature generation
-    const { bls12_381 } = await import("@noble/curves/bls12-381.js");
-    const messageWithDomain = new Uint8Array([
-      ...new TextEncoder().encode("SUI-MULTI-IBS-V1"),
-      ...messageBytes,
-    ]);
-
-    // Create a deterministic signature using real cryptographic operations
-    // In a full implementation, this would use actual derived keys from Seal
-    const hash = await crypto.subtle.digest("SHA-256", messageWithDomain);
-
-    // Generate a deterministic scalar for signature
-    const keyIdsCombined = keyServerIds.slice(0, threshold).join("");
-    const keyHash = await crypto.subtle.digest("SHA-256", new TextEncoder().encode(keyIdsCombined));
-
-    // Combine hashes for signature determinism
-    const combined = new Uint8Array(hash.byteLength + keyHash.byteLength);
-    combined.set(new Uint8Array(hash), 0);
-    combined.set(new Uint8Array(keyHash), hash.byteLength);
-
-    const finalHash = await crypto.subtle.digest("SHA-256", combined);
-
-    // Create proper G1 signature using BLS12-381
-    const hashArray = new Uint8Array(finalHash);
-    const scalar = hashArray.slice(0, 32); // Use first 32 bytes as scalar
-
-    // Generate G1 point: scalar * G1.GENERATOR
-    const g1Generator = bls12_381.G1.Point.BASE;
-    const scalarBigInt = BigInt(
-      `0x${Array.from(scalar)
-        .map((b) => b.toString(16).padStart(2, "0"))
-        .join("")}`,
-    );
-    // Reduce modulo curve order to ensure valid scalar
-    const reducedScalar = scalarBigInt % bls12_381.fields.Fr.ORDER;
-    const signaturePoint = g1Generator.multiply(reducedScalar);
-
-    // Convert to compressed bytes (48 bytes for G1)
-    const signature = signaturePoint.toBytes(true);
-
-    return {
-      signature_g1: Array.from(signature),
-      message: Array.from(messageBytes),
-    };
-  } catch (error) {
-    throw new Error(`Real BLS signature generation failed: ${error}`);
-  }
-}
-
-/**
- * Clear the public key cache (useful for tests)
- */
-export function clearPublicKeyCache(): void {
-  publicKeyCache.clear();
-}
