--- sui/scripts/multi-ibs-with-seal.ts
@@ -1,456 +1,0 @@
-#!/usr/bin/env tsx
-
-/**
- * Multi-IBS with Seal SDK Integration
- *
- * This script demonstrates how to integrate actual Seal Key Servers
- * with Multi-IBS signature aggregation using:
- * 1. Seal SDK for sk_ID_i retrieval from Key Servers
- * 2. @noble/curves for BLS signature operations
- * 3. G1Element aggregation from multiple Key Servers
- * 4. Real cryptographic Multi-IBS implementation
- *
- * ===== KEY CONCEPT: sk_ID_i as G1Element =====
- *
- * What is sk_ID_i?
- * - Identity-Based Encryption (IBE) secret key for identity ID
- * - Generated by Seal Key Server i: sk_ID_i = H₁(ID) × msk_i
- * - H₁ is hash-to-G1, msk_i is master secret key of server i
- * - sk_ID_i is a G1Element (48 bytes compressed)
- *
- * Why can we use IBE keys for signatures?
- * - IBE secret keys are mathematically compatible with BLS signatures
- * - sk_ID = Σ sk_ID_i enables threshold signature capability
- * - Each sk_ID_i can independently create valid signatures
- * - Aggregation maintains security properties
- *
- * Mock vs Real implementation:
- * - This script includes both mock (for demo) and real Seal integration
- * - Real usage would uncomment the Seal SDK calls
- * - Mock keys are deterministic for testing consistency
- */
-
-import { SealClient, SessionKey } from "@mysten/seal";
-import { getFullnodeUrl, SuiClient } from "@mysten/sui/client";
-import { Ed25519Keypair } from "@mysten/sui/keypairs/ed25519";
-import { Transaction } from "@mysten/sui/transactions";
-import { bls12_381 } from "@noble/curves/bls12-381.js";
-import { counterPackage } from "../src/abi/index.js";
-
-/**
- * CRITICAL: Same signature scheme choice as test file
- *
- * We use shortSignatures (G1 signatures, G2 public keys) for:
- * - 48-byte signatures (efficient on-chain storage)
- * - 96-byte public keys (aggregated once, stored off-chain in Key Servers)
- * - Consistency with Move contract expectations
- * - Optimal gas costs for repeated signature verification
- */
-const blss = bls12_381.shortSignatures;
-
-// Configuration
-const NETWORK = "testnet";
-const COUNTER_PACKAGE_ID = counterPackage.packageId;
-/**
- * THRESHOLD SIGNATURE CONFIGURATION
- *
- * 2-of-3 threshold means:
- * - Any 2 Key Servers can cooperate to create valid signatures
- * - Single Key Server compromise doesn't break security
- * - Provides fault tolerance (1 server can be offline)
- * - Balances security vs availability
- *
- * Higher thresholds (e.g., 3-of-5) would provide more security
- * but reduce availability and increase coordination complexity.
- */
-const THRESHOLD = 2; // 2-of-3 threshold
-
-/**
- * Seal Key Server configurations (Testnet servers)
- *
- * These are real Key Server object IDs on Sui testnet:
- * - Each server has its own master secret key (msk_i)
- * - Servers cooperate to provide threshold IBE service
- * - Object IDs are publicly known, but master keys are kept secret
- * - Each server independently computes sk_ID_i for given identity ID
- */
-const _KEY_SERVERS = [
-  {
-    name: "Studio Mirai",
-    objectId: "0x164ac3d2b3b8694b8181c13f671950004765c23f270321a45fdd04d40cccf0f2",
-  },
-  {
-    name: "Ruby Node",
-    objectId: "0x6068c0acb197dddbacd4746a9de7f025b2ed5a5b6c1b1ab44dade4426d141da2",
-  },
-  {
-    name: "NodeInfra",
-    objectId: "0x5466b7df5c15b508678d51496ada8afab0d6f70a01c10613123382b1b8131007",
-  },
-];
-
-interface KeyShare {
-  serverIndex: number;
-  serverId: string;
-  /**
-   * sk_ID_i as G1Element bytes (typically 48 bytes compressed)
-   *
-   * This is the IBE secret key from Key Server i for identity ID:
-   * - Generated by: sk_ID_i = H₁(ID) × msk_i
-   * - H₁: hash-to-G1 function (deterministic)
-   * - msk_i: master secret key of Key Server i (kept secret)
-   * - Result: G1 curve point that can be used for BLS signatures
-   *
-   * Security properties:
-   * - Each server can independently compute this
-   * - Servers never need to communicate with each other
-   * - Aggregation happens client-side: sk_ID = Σ sk_ID_i
-   */
-  secretKey: Uint8Array; // sk_ID_i as G1Element bytes
-}
-
-interface MultiIBSSignature {
-  signature: Uint8Array; // G1 signature (48 bytes)
-  message: Uint8Array;
-  identity: string;
-}
-
-class SealMultiIBSAggregator {
-  private suiClient: SuiClient;
-  private sealClient: SealClient;
-
-  constructor(network: "testnet" | "mainnet" = NETWORK) {
-    this.suiClient = new SuiClient({ url: getFullnodeUrl(network) });
-
-    // Initialize SealClient for real Key Server operations
-    this.sealClient = new SealClient({
-      networkConfig: network, // Use the passed network parameter
-      suiClient: this.suiClient,
-      serverConfigs: [
-        {
-          objectId: "0x164ac3d2b3b8694b8181c13f671950004765c23f270321a45fdd04d40cccf0f2",
-          url: "https://seal.studio-mirai.com", // Studio Mirai
-          weight: 1,
-        },
-        {
-          objectId: "0x6068c0acb197dddbacd4746a9de7f025b2ed5a5b6c1b1ab44dade4426d141da2",
-          url: "https://seal.ruby-node.com", // Ruby Node
-          weight: 1,
-        },
-        {
-          objectId: "0x5466b7df5c15b508678d51496ada8afab0d6f70a01c10613123382b1b8131007",
-          url: "https://seal.nodeinfra.com", // NodeInfra
-          weight: 1,
-        },
-      ],
-    });
-  }
-
-  /**
-   * Retrieve sk_ID_i from Seal Key Servers for a specific message
-   * @param counterId Multi-IBS counter object ID
-   * @param signerKeypair Ed25519 keypair for signing
-   * @param message The message to sign
-   * @param requiredCount Number of key shares needed
-   */
-  async fetchSecretKeyShares(
-    counterId: string,
-    signerKeypair: Ed25519Keypair,
-    message: string,
-    requiredCount: number = THRESHOLD,
-  ): Promise<KeyShare[]> {
-    /**
-     * REAL SEAL SDK INTEGRATION - CORRECT LIFECYCLE
-     *
-     * Seal flow (following dpp-pilot pattern):
-     * 1. Construct InnerID: counter_id || signer_address
-     * 2. Create SessionKey FIRST (before seal_approve)
-     * 3. Build seal_approve PTB (don't execute, just build for txBytes)
-     * 4. Fetch IBE keys from Key Servers using SessionKey and txBytes
-     */
-
-    try {
-      // Step 1: Construct IBE Identity = counter_id || signer_address || message
-      const signerAddress = signerKeypair.getPublicKey().toSuiAddress();
-      const messageBytes = new TextEncoder().encode(message);
-
-      // Prepare message with domain separation matching Move contract
-      const messageWithDomain = new Uint8Array([
-        ...new TextEncoder().encode("SUI-MULTI-IBS-V1"),
-        ...messageBytes,
-      ]);
-
-      // Convert counter ID to bytes (same as Move object::id().to_bytes())
-      const counterIdBytes = Array.from(Buffer.from(counterId.replace("0x", ""), "hex"));
-
-      // Convert signer address to bytes (same as Move bcs::to_bytes(&address))
-      const signerBytes = Array.from(Buffer.from(signerAddress.replace("0x", ""), "hex"));
-
-      // Concatenate: counter_id || signer_address || message_with_domain
-      // This ensures IBE key is specific to this message
-      const ibeIdBytes = [...counterIdBytes, ...signerBytes, ...Array.from(messageWithDomain)];
-
-      // Convert to hex string for Seal SDK
-      const ibeIdHex = `0x${Buffer.from(ibeIdBytes).toString("hex")}`;
-
-      const sessionKey = await SessionKey.create({
-        address: signerAddress,
-        packageId: COUNTER_PACKAGE_ID,
-        ttlMin: 10,
-        signer: signerKeypair,
-        suiClient: this.suiClient,
-      });
-
-      // Step 3: Build seal_approve transaction for txBytes (DO NOT EXECUTE)
-      // This transaction is only for generating txBytes needed by getDerivedKeys
-      // We don't execute it to avoid changing Key Server object versions
-      const approveTx = new Transaction();
-      approveTx.moveCall({
-        target: `${COUNTER_PACKAGE_ID}::multi_ibs_counter::seal_approve`,
-        arguments: [
-          approveTx.pure.vector("u8", ibeIdBytes), // Use the unified IBE ID
-          approveTx.object(counterId),
-          approveTx.pure.vector("u8", Array.from(messageBytes)), // Add message parameter
-        ],
-      });
-
-      // Generate txBytes without executing the transaction
-      // This avoids object version conflicts in tests
-      const txBytes = await approveTx.build({
-        client: this.suiClient,
-        onlyTransactionKind: true,
-      });
-
-      // Step 4: Get IBE derived keys from Key Servers using getDerivedKeys
-      // getDerivedKeys returns Map<string, G1Element> where key is server objectId
-      // Use the message-specific IBE ID for key derivation
-      const derivedKeys = await this.sealClient.getDerivedKeys({
-        id: ibeIdHex, // Use hex-encoded IBE ID (includes message)
-        sessionKey,
-        txBytes,
-        threshold: requiredCount,
-      });
-
-      // Check if derivedKeys is a Map
-      if (!derivedKeys || typeof derivedKeys.size !== "number") {
-        throw new Error(`getDerivedKeys returned unexpected result: ${derivedKeys}`);
-      }
-
-      // Step 5: Convert Map to KeyShare format
-      const keyShares: KeyShare[] = [];
-      let serverIndex = 0;
-
-      for (const [serverId, derivedKey] of derivedKeys) {
-        const keyBytes = derivedKey.key.toBytes();
-
-        keyShares.push({
-          serverIndex,
-          serverId,
-          secretKey: keyBytes, // BonehFranklinBLS12381DerivedKey.key.toBytes() -> Uint8Array (48 bytes)
-        });
-        serverIndex++;
-      }
-
-      return keyShares;
-    } catch (error) {
-      // Provide more specific error context
-      const errorMessage = error?.message || String(error) || "Unknown error";
-      if (errorMessage.includes("notExists")) {
-      } else if (errorMessage.includes("NoAccess")) {
-      }
-
-      throw new Error(
-        `Real Key Server integration failed: ${errorMessage}. Mock fallback is forbidden.`,
-      );
-    }
-  }
-
-  /**
-   * Aggregate IBE secret keys (G1 elements) directly as signature
-   * In IBE-based BLS, aggregated IBE keys ARE the signature for the identity
-   */
-  aggregateIBESignature(keyShares: KeyShare[]): Uint8Array {
-    if (keyShares.length === 0) {
-      throw new Error("No key shares to aggregate");
-    }
-
-    // Parse first IBE key as G1 element and aggregate others
-    let aggregatedSignature = bls12_381.G1.Point.fromBytes(keyShares[0].secretKey);
-
-    for (let i = 1; i < keyShares.length; i++) {
-      const ibeKey = bls12_381.G1.Point.fromBytes(keyShares[i].secretKey);
-      aggregatedSignature = aggregatedSignature.add(ibeKey);
-    }
-
-    const result = aggregatedSignature.toBytes(true); // 48 bytes compressed G1 point
-    return result;
-  }
-
-  /**
-   * Legacy method for compatibility - now delegates to IBE signature aggregation
-   */
-  aggregateSecretKeys(keyShares: KeyShare[]): Uint8Array {
-    // IBE keys from Seal SDK are G1 elements, not scalars
-    // We need to return them as-is for direct signature use
-    return this.aggregateIBESignature(keyShares);
-  }
-
-  /**
-   * Create Multi-IBS signature using aggregated IBE signature
-   * In IBE-based BLS, the aggregated IBE keys ARE the signature
-   */
-  createMultiIBSSignature(
-    aggregatedIBESignature: Uint8Array,
-    message: string,
-    identity: string,
-  ): MultiIBSSignature {
-    const messageBytes = new TextEncoder().encode(message);
-
-    // Create Multi-IBS signature using aggregated IBE signature
-
-    // For IBE-based signatures, the aggregated IBE key IS the signature
-    // No additional signing operation needed
-    return {
-      signature: aggregatedIBESignature, // Use IBE signature directly
-      message: messageBytes,
-      identity,
-    };
-  }
-
-  /**
-   * Verify Multi-IBS signature before submission
-   */
-  verifyMultiIBSSignature(multiSig: MultiIBSSignature, aggregatedPublicKey: Uint8Array): boolean {
-    // Use same domain separation as Move contract
-    const messageWithDomain = new Uint8Array([
-      ...new TextEncoder().encode("SUI-MULTI-IBS-V1"),
-      ...multiSig.message,
-    ]);
-    const DST = "BLS_SIG_BLS12381G1_XMD:SHA-256_SSWU_RO_NUL_";
-    const hashedMessage = blss.hash(messageWithDomain, DST);
-
-    // Convert bytes back to Point objects for verification
-    const sigPoint = bls12_381.G1.Point.fromBytes(multiSig.signature);
-    const pubKeyPoint = bls12_381.G2.Point.fromBytes(aggregatedPublicKey);
-
-    return blss.verify(sigPoint, hashedMessage, pubKeyPoint);
-  }
-
-  /**
-   * Submit Multi-IBS signature to blockchain for verification
-   */
-  async submitMultiIBSSignature(
-    counterId: string,
-    multiSig: MultiIBSSignature,
-    signerKeypair: Ed25519Keypair,
-  ): Promise<string> {
-    const tx = new Transaction();
-
-    // Call verify_and_create_proof with aggregated signature
-    const proof = tx.moveCall({
-      target: `${COUNTER_PACKAGE_ID}::multi_ibs_counter::verify_and_create_proof`,
-      arguments: [
-        tx.object(counterId),
-        tx.pure.vector("u8", Array.from(multiSig.signature)),
-        tx.pure.vector("u8", Array.from(multiSig.message)),
-      ],
-    });
-
-    // Use proof to increment counter
-    tx.moveCall({
-      target: `${COUNTER_PACKAGE_ID}::multi_ibs_counter::increment`,
-      arguments: [tx.object(counterId), proof],
-    });
-
-    const result = await this.suiClient.signAndExecuteTransaction({
-      signer: signerKeypair,
-      transaction: tx,
-      options: { showEffects: true, showObjectChanges: true },
-    });
-
-    if (result.effects?.status?.status !== "success") {
-      throw new Error(`Transaction failed: ${result.effects?.status?.error}`);
-    }
-
-    return result.digest;
-  }
-
-  /**
-   * Get aggregated public key from Key Servers using Seal SDK
-   * @param keyServerIds Array of Key Server object IDs to fetch public keys from
-   */
-  async getAggregatedPublicKey(keyServerIds: string[]): Promise<Uint8Array> {
-    try {
-      // Get public keys from Seal Key Servers
-      const publicKeys = await this.sealClient.getPublicKeys(keyServerIds);
-
-      // Aggregate G2 public keys using curve point addition
-      let aggregatedPK = bls12_381.G2.Point.ZERO; // Identity element (neutral element for addition)
-
-      for (const publicKey of publicKeys) {
-        // Convert each public key to G2 point and add to aggregate
-        const g2Point = bls12_381.G2.Point.fromBytes(publicKey.toBytes());
-        aggregatedPK = aggregatedPK.add(g2Point);
-      }
-      return aggregatedPK.toBytes(); // Return as 96-byte compressed G2 point
-    } catch (error) {
-      throw new Error(`Failed to aggregate public keys: ${error}`);
-    }
-  }
-
-  /**
-   * Get public key from secret key
-   */
-  getPublicKey(secretKey: Uint8Array): Uint8Array {
-    return blss.getPublicKey(secretKey).toBytes();
-  }
-
-  /**
-   * Get Seal Shard public key for a specific Key Server
-   * Used by real-seal-keys.ts utility
-   */
-  async getSealShardPublicKey(keyServerId: string): Promise<Uint8Array> {
-    try {
-      // Get public key from specific Seal Key Server
-      const publicKeys = await this.sealClient.getPublicKeys([keyServerId]);
-
-      if (!publicKeys || publicKeys.length === 0) {
-        throw new Error(`No public key found for Key Server: ${keyServerId}`);
-      }
-
-      return publicKeys[0].toBytes();
-    } catch (error) {
-      throw new Error(`Failed to get public key for ${keyServerId}: ${error}`);
-    }
-  }
-}
-
-// Demo function
-async function demonstrateMultiIBS() {
-  const aggregator = new SealMultiIBSAggregator();
-  const signerKeypair = Ed25519Keypair.generate();
-
-  try {
-    // Step 1: Fetch IBE key shares from Seal servers
-    const identity = signerKeypair.getPublicKey().toSuiAddress();
-    const message = `increment-counter-${Date.now()}`;
-    const keyShares = await aggregator.fetchSecretKeyShares(identity, signerKeypair, message, 2);
-    const aggregatedSK = aggregator.aggregateSecretKeys(keyShares);
-
-    // Step 2: Create Multi-IBS signature
-    const _multiSig = aggregator.createMultiIBSSignature(aggregatedSK, message, identity);
-
-    // Note: Blockchain submission would require deployed contract
-    // const txDigest = await aggregator.submitMultiIBSSignature(counterId, multiSig, signerKeypair);
-    // console.log(`📝 Transaction submitted: ${txDigest}`);
-  } catch (_error) {
-    process.exit(1);
-  }
-}
-
-// Execute demo if run directly
-if (import.meta.url === `file://${process.argv[1]}`) {
-  demonstrateMultiIBS();
-}
-
-export { SealMultiIBSAggregator };
