+++ sui/src/abi/counter.abi.ts
@@ -3,7 +3,7 @@
  * 自動生成されたファイル - 手動で編集しないでください
  *
  * Generated by: scripts/suigenv0.ts
- * Generated at: 2025-09-27T03:52:05.321Z
+ * Generated at: 2025-09-27T08:26:57.472Z
  */
 
 import {
@@ -13,13 +13,13 @@ import {
 } from "@mysten/sui/transactions";
 
 export const counterAbi = {
-  multi_ibs_counter: {
+  bls_multisig_counter: {
     fileFormatVersion: 6,
-    address: "0x6ce277618bf01afd77ef1de3ff19dda6df713387aaf91d53f6a58c5904b5ec21",
-    name: "multi_ibs_counter",
+    address: "0x9bd24ba5fd617512a16ae57fbd84dedbe54d3389e3a227311d16c3159ebcc036",
+    name: "bls_multisig_counter",
     friends: [],
     structs: {
-      AggregatedPublicKey: {
+      BlsAggregatedPk: {
         abilities: {
           abilities: ["Key"],
         },
@@ -57,6 +57,12 @@ export const counterAbi = {
             },
           },
           {
+            name: "public_key_bytes",
+            type: {
+              Vector: "U8",
+            },
+          },
+          {
             name: "counter_id",
             type: {
               Struct: {
@@ -86,7 +92,7 @@ export const counterAbi = {
           },
         ],
       },
-      MultiIBSConfig: {
+      BlsMultisigConfig: {
         abilities: {
           abilities: ["Store"],
         },
@@ -135,7 +141,7 @@ export const counterAbi = {
           },
         ],
       },
-      MultiIBSCounter: {
+      BlsMultisigCounter: {
         abilities: {
           abilities: ["Key"],
         },
@@ -160,16 +166,16 @@ export const counterAbi = {
             name: "config",
             type: {
               Struct: {
-                address: "0x6ce277618bf01afd77ef1de3ff19dda6df713387aaf91d53f6a58c5904b5ec21",
-                module: "multi_ibs_counter",
-                name: "MultiIBSConfig",
+                address: "0x9bd24ba5fd617512a16ae57fbd84dedbe54d3389e3a227311d16c3159ebcc036",
+                module: "bls_multisig_counter",
+                name: "BlsMultisigConfig",
                 typeArguments: [],
               },
             },
           },
         ],
       },
-      MultiIBSProof: {
+      BlsMultisigProof: {
         abilities: {
           abilities: ["Key"],
         },
@@ -225,9 +231,9 @@ export const counterAbi = {
           {
             Reference: {
               Struct: {
-                address: "0x6ce277618bf01afd77ef1de3ff19dda6df713387aaf91d53f6a58c5904b5ec21",
-                module: "multi_ibs_counter",
-                name: "MultiIBSCounter",
+                address: "0x9bd24ba5fd617512a16ae57fbd84dedbe54d3389e3a227311d16c3159ebcc036",
+                module: "bls_multisig_counter",
+                name: "BlsMultisigCounter",
                 typeArguments: [],
               },
             },
@@ -235,9 +241,9 @@ export const counterAbi = {
           {
             MutableReference: {
               Struct: {
-                address: "0x6ce277618bf01afd77ef1de3ff19dda6df713387aaf91d53f6a58c5904b5ec21",
-                module: "multi_ibs_counter",
-                name: "AggregatedPublicKey",
+                address: "0x9bd24ba5fd617512a16ae57fbd84dedbe54d3389e3a227311d16c3159ebcc036",
+                module: "bls_multisig_counter",
+                name: "BlsAggregatedPk",
                 typeArguments: [],
               },
             },
@@ -261,9 +267,9 @@ export const counterAbi = {
           {
             Reference: {
               Struct: {
-                address: "0x6ce277618bf01afd77ef1de3ff19dda6df713387aaf91d53f6a58c5904b5ec21",
-                module: "multi_ibs_counter",
-                name: "AggregatedPublicKey",
+                address: "0x9bd24ba5fd617512a16ae57fbd84dedbe54d3389e3a227311d16c3159ebcc036",
+                module: "bls_multisig_counter",
+                name: "BlsAggregatedPk",
                 typeArguments: [],
               },
             },
@@ -280,16 +286,16 @@ export const counterAbi = {
           },
         ],
       },
-      destroy_aggregated_public_key: {
+      destroy_bls_aggregated_pk: {
         visibility: "Public",
         isEntry: false,
         typeParameters: [],
         parameters: [
           {
             Struct: {
-              address: "0x6ce277618bf01afd77ef1de3ff19dda6df713387aaf91d53f6a58c5904b5ec21",
-              module: "multi_ibs_counter",
-              name: "AggregatedPublicKey",
+              address: "0x9bd24ba5fd617512a16ae57fbd84dedbe54d3389e3a227311d16c3159ebcc036",
+              module: "bls_multisig_counter",
+              name: "BlsAggregatedPk",
               typeArguments: [],
             },
           },
@@ -315,9 +321,9 @@ export const counterAbi = {
           {
             Reference: {
               Struct: {
-                address: "0x6ce277618bf01afd77ef1de3ff19dda6df713387aaf91d53f6a58c5904b5ec21",
-                module: "multi_ibs_counter",
-                name: "AggregatedPublicKey",
+                address: "0x9bd24ba5fd617512a16ae57fbd84dedbe54d3389e3a227311d16c3159ebcc036",
+                module: "bls_multisig_counter",
+                name: "BlsAggregatedPk",
                 typeArguments: [],
               },
             },
@@ -333,9 +339,9 @@ export const counterAbi = {
           {
             Reference: {
               Struct: {
-                address: "0x6ce277618bf01afd77ef1de3ff19dda6df713387aaf91d53f6a58c5904b5ec21",
-                module: "multi_ibs_counter",
-                name: "AggregatedPublicKey",
+                address: "0x9bd24ba5fd617512a16ae57fbd84dedbe54d3389e3a227311d16c3159ebcc036",
+                module: "bls_multisig_counter",
+                name: "BlsAggregatedPk",
                 typeArguments: [],
               },
             },
@@ -364,25 +370,25 @@ export const counterAbi = {
           {
             MutableReference: {
               Struct: {
-                address: "0x6ce277618bf01afd77ef1de3ff19dda6df713387aaf91d53f6a58c5904b5ec21",
-                module: "multi_ibs_counter",
-                name: "MultiIBSCounter",
+                address: "0x9bd24ba5fd617512a16ae57fbd84dedbe54d3389e3a227311d16c3159ebcc036",
+                module: "bls_multisig_counter",
+                name: "BlsMultisigCounter",
                 typeArguments: [],
               },
             },
           },
           {
             Struct: {
-              address: "0x6ce277618bf01afd77ef1de3ff19dda6df713387aaf91d53f6a58c5904b5ec21",
-              module: "multi_ibs_counter",
-              name: "MultiIBSProof",
+              address: "0x9bd24ba5fd617512a16ae57fbd84dedbe54d3389e3a227311d16c3159ebcc036",
+              module: "bls_multisig_counter",
+              name: "BlsMultisigProof",
               typeArguments: [],
             },
           },
         ],
         return: [],
       },
-      new_aggregated_public_key: {
+      new_bls_aggregated_pk: {
         visibility: "Public",
         isEntry: false,
         typeParameters: [],
@@ -390,9 +396,9 @@ export const counterAbi = {
           {
             Reference: {
               Struct: {
-                address: "0x6ce277618bf01afd77ef1de3ff19dda6df713387aaf91d53f6a58c5904b5ec21",
-                module: "multi_ibs_counter",
-                name: "MultiIBSCounter",
+                address: "0x9bd24ba5fd617512a16ae57fbd84dedbe54d3389e3a227311d16c3159ebcc036",
+                module: "bls_multisig_counter",
+                name: "BlsMultisigCounter",
                 typeArguments: [],
               },
             },
@@ -411,9 +417,9 @@ export const counterAbi = {
         return: [
           {
             Struct: {
-              address: "0x6ce277618bf01afd77ef1de3ff19dda6df713387aaf91d53f6a58c5904b5ec21",
-              module: "multi_ibs_counter",
-              name: "AggregatedPublicKey",
+              address: "0x9bd24ba5fd617512a16ae57fbd84dedbe54d3389e3a227311d16c3159ebcc036",
+              module: "bls_multisig_counter",
+              name: "BlsAggregatedPk",
               typeArguments: [],
             },
           },
@@ -427,9 +433,9 @@ export const counterAbi = {
           {
             Reference: {
               Struct: {
-                address: "0x6ce277618bf01afd77ef1de3ff19dda6df713387aaf91d53f6a58c5904b5ec21",
-                module: "multi_ibs_counter",
-                name: "AggregatedPublicKey",
+                address: "0x9bd24ba5fd617512a16ae57fbd84dedbe54d3389e3a227311d16c3159ebcc036",
+                module: "bls_multisig_counter",
+                name: "BlsAggregatedPk",
                 typeArguments: [],
               },
             },
@@ -468,9 +474,9 @@ export const counterAbi = {
           {
             Reference: {
               Struct: {
-                address: "0x6ce277618bf01afd77ef1de3ff19dda6df713387aaf91d53f6a58c5904b5ec21",
-                module: "multi_ibs_counter",
-                name: "MultiIBSCounter",
+                address: "0x9bd24ba5fd617512a16ae57fbd84dedbe54d3389e3a227311d16c3159ebcc036",
+                module: "bls_multisig_counter",
+                name: "BlsMultisigCounter",
                 typeArguments: [],
               },
             },
@@ -499,9 +505,9 @@ export const counterAbi = {
           {
             Reference: {
               Struct: {
-                address: "0x6ce277618bf01afd77ef1de3ff19dda6df713387aaf91d53f6a58c5904b5ec21",
-                module: "multi_ibs_counter",
-                name: "MultiIBSCounter",
+                address: "0x9bd24ba5fd617512a16ae57fbd84dedbe54d3389e3a227311d16c3159ebcc036",
+                module: "bls_multisig_counter",
+                name: "BlsMultisigCounter",
                 typeArguments: [],
               },
             },
@@ -517,9 +523,9 @@ export const counterAbi = {
           {
             Reference: {
               Struct: {
-                address: "0x6ce277618bf01afd77ef1de3ff19dda6df713387aaf91d53f6a58c5904b5ec21",
-                module: "multi_ibs_counter",
-                name: "MultiIBSCounter",
+                address: "0x9bd24ba5fd617512a16ae57fbd84dedbe54d3389e3a227311d16c3159ebcc036",
+                module: "bls_multisig_counter",
+                name: "BlsMultisigCounter",
                 typeArguments: [],
               },
             },
@@ -606,9 +612,9 @@ export const counterAbi = {
           {
             Reference: {
               Struct: {
-                address: "0x6ce277618bf01afd77ef1de3ff19dda6df713387aaf91d53f6a58c5904b5ec21",
-                module: "multi_ibs_counter",
-                name: "MultiIBSCounter",
+                address: "0x9bd24ba5fd617512a16ae57fbd84dedbe54d3389e3a227311d16c3159ebcc036",
+                module: "bls_multisig_counter",
+                name: "BlsMultisigCounter",
                 typeArguments: [],
               },
             },
@@ -624,9 +630,9 @@ export const counterAbi = {
           {
             Reference: {
               Struct: {
-                address: "0x6ce277618bf01afd77ef1de3ff19dda6df713387aaf91d53f6a58c5904b5ec21",
-                module: "multi_ibs_counter",
-                name: "MultiIBSCounter",
+                address: "0x9bd24ba5fd617512a16ae57fbd84dedbe54d3389e3a227311d16c3159ebcc036",
+                module: "bls_multisig_counter",
+                name: "BlsMultisigCounter",
                 typeArguments: [],
               },
             },
@@ -642,9 +648,9 @@ export const counterAbi = {
           {
             Reference: {
               Struct: {
-                address: "0x6ce277618bf01afd77ef1de3ff19dda6df713387aaf91d53f6a58c5904b5ec21",
-                module: "multi_ibs_counter",
-                name: "MultiIBSCounter",
+                address: "0x9bd24ba5fd617512a16ae57fbd84dedbe54d3389e3a227311d16c3159ebcc036",
+                module: "bls_multisig_counter",
+                name: "BlsMultisigCounter",
                 typeArguments: [],
               },
             },
@@ -652,9 +658,9 @@ export const counterAbi = {
           {
             Reference: {
               Struct: {
-                address: "0x6ce277618bf01afd77ef1de3ff19dda6df713387aaf91d53f6a58c5904b5ec21",
-                module: "multi_ibs_counter",
-                name: "AggregatedPublicKey",
+                address: "0x9bd24ba5fd617512a16ae57fbd84dedbe54d3389e3a227311d16c3159ebcc036",
+                module: "bls_multisig_counter",
+                name: "BlsAggregatedPk",
                 typeArguments: [],
               },
             },
@@ -679,9 +685,9 @@ export const counterAbi = {
         return: [
           {
             Struct: {
-              address: "0x6ce277618bf01afd77ef1de3ff19dda6df713387aaf91d53f6a58c5904b5ec21",
-              module: "multi_ibs_counter",
-              name: "MultiIBSProof",
+              address: "0x9bd24ba5fd617512a16ae57fbd84dedbe54d3389e3a227311d16c3159ebcc036",
+              module: "bls_multisig_counter",
+              name: "BlsMultisigProof",
               typeArguments: [],
             },
           },
@@ -691,7 +697,7 @@ export const counterAbi = {
   },
   owned_counter: {
     fileFormatVersion: 6,
-    address: "0x6ce277618bf01afd77ef1de3ff19dda6df713387aaf91d53f6a58c5904b5ec21",
+    address: "0x9bd24ba5fd617512a16ae57fbd84dedbe54d3389e3a227311d16c3159ebcc036",
     name: "owned_counter",
     friends: [],
     structs: {
@@ -728,7 +734,7 @@ export const counterAbi = {
           {
             MutableReference: {
               Struct: {
-                address: "0x6ce277618bf01afd77ef1de3ff19dda6df713387aaf91d53f6a58c5904b5ec21",
+                address: "0x9bd24ba5fd617512a16ae57fbd84dedbe54d3389e3a227311d16c3159ebcc036",
                 module: "owned_counter",
                 name: "OwnedCounter",
                 typeArguments: [],
@@ -757,7 +763,7 @@ export const counterAbi = {
         return: [
           {
             Struct: {
-              address: "0x6ce277618bf01afd77ef1de3ff19dda6df713387aaf91d53f6a58c5904b5ec21",
+              address: "0x9bd24ba5fd617512a16ae57fbd84dedbe54d3389e3a227311d16c3159ebcc036",
               module: "owned_counter",
               name: "OwnedCounter",
               typeArguments: [],
@@ -773,7 +779,7 @@ export const counterAbi = {
           {
             MutableReference: {
               Struct: {
-                address: "0x6ce277618bf01afd77ef1de3ff19dda6df713387aaf91d53f6a58c5904b5ec21",
+                address: "0x9bd24ba5fd617512a16ae57fbd84dedbe54d3389e3a227311d16c3159ebcc036",
                 module: "owned_counter",
                 name: "OwnedCounter",
                 typeArguments: [],
@@ -792,7 +798,7 @@ export const counterAbi = {
           {
             Reference: {
               Struct: {
-                address: "0x6ce277618bf01afd77ef1de3ff19dda6df713387aaf91d53f6a58c5904b5ec21",
+                address: "0x9bd24ba5fd617512a16ae57fbd84dedbe54d3389e3a227311d16c3159ebcc036",
                 module: "owned_counter",
                 name: "OwnedCounter",
                 typeArguments: [],
@@ -806,7 +812,7 @@ export const counterAbi = {
   },
   shared_counter: {
     fileFormatVersion: 6,
-    address: "0x6ce277618bf01afd77ef1de3ff19dda6df713387aaf91d53f6a58c5904b5ec21",
+    address: "0x9bd24ba5fd617512a16ae57fbd84dedbe54d3389e3a227311d16c3159ebcc036",
     name: "shared_counter",
     friends: [],
     structs: {
@@ -843,7 +849,7 @@ export const counterAbi = {
           {
             MutableReference: {
               Struct: {
-                address: "0x6ce277618bf01afd77ef1de3ff19dda6df713387aaf91d53f6a58c5904b5ec21",
+                address: "0x9bd24ba5fd617512a16ae57fbd84dedbe54d3389e3a227311d16c3159ebcc036",
                 module: "shared_counter",
                 name: "SharedCounter",
                 typeArguments: [],
@@ -861,7 +867,7 @@ export const counterAbi = {
           {
             MutableReference: {
               Struct: {
-                address: "0x6ce277618bf01afd77ef1de3ff19dda6df713387aaf91d53f6a58c5904b5ec21",
+                address: "0x9bd24ba5fd617512a16ae57fbd84dedbe54d3389e3a227311d16c3159ebcc036",
                 module: "shared_counter",
                 name: "SharedCounter",
                 typeArguments: [],
@@ -898,7 +904,7 @@ export const counterAbi = {
           {
             Reference: {
               Struct: {
-                address: "0x6ce277618bf01afd77ef1de3ff19dda6df713387aaf91d53f6a58c5904b5ec21",
+                address: "0x9bd24ba5fd617512a16ae57fbd84dedbe54d3389e3a227311d16c3159ebcc036",
                 module: "shared_counter",
                 name: "SharedCounter",
                 typeArguments: [],
@@ -915,32 +921,33 @@ export const counterAbi = {
 export type CounterModules = typeof counterAbi;
 
 // Generated TypeScript interfaces
-export interface Multi_ibs_counterAggregatedPublicKeyType {
+export interface Bls_multisig_counterBlsAggregatedPkType {
   id: { id: string };
   public_key_g2: unknown;
+  public_key_bytes: Uint8Array;
   counter_id: unknown;
   seal_shard_count: string;
   included_seal_shard_ids: unknown[];
 }
 
-export interface Multi_ibs_counterMultiIBSConfigType {
+export interface Bls_multisig_counterBlsMultisigConfigType {
   seal_shard_table: { id: string };
   threshold: string;
 }
 
-export interface Multi_ibs_counterMultiIBSCounterType {
+export interface Bls_multisig_counterBlsMultisigCounterType {
   id: { id: string };
   value: string;
   config: unknown;
 }
 
-export interface Multi_ibs_counterMultiIBSProofType {
+export interface Bls_multisig_counterBlsMultisigProofType {
   id: { id: string };
   counter_id: unknown;
   verified_signer_count: string;
 }
 
-export interface Multi_ibs_counterPackageMarkerType {
+export interface Bls_multisig_counterPackageMarkerType {
   dummy_field: boolean;
 }
 
@@ -955,9 +962,9 @@ export interface Shared_counterSharedCounterType {
 }
 
 // Generated type guards
-export function isMulti_ibs_counterAggregatedPublicKeyType(
+export function isBls_multisig_counterBlsAggregatedPkType(
   data: unknown,
-): data is Multi_ibs_counterAggregatedPublicKeyType {
+): data is Bls_multisig_counterBlsAggregatedPkType {
   if (!data || typeof data !== "object") return false;
   const obj = data as Record<string, unknown>;
 
@@ -965,6 +972,7 @@ export function isMulti_ibs_counterAggregatedPublicKeyType(
   const requiredFields = [
     "id",
     "public_key_g2",
+    "public_key_bytes",
     "counter_id",
     "seal_shard_count",
     "included_seal_shard_ids",
@@ -982,9 +990,9 @@ export function isMulti_ibs_counterAggregatedPublicKeyType(
   return true;
 }
 
-export function isMulti_ibs_counterMultiIBSConfigType(
+export function isBls_multisig_counterBlsMultisigConfigType(
   data: unknown,
-): data is Multi_ibs_counterMultiIBSConfigType {
+): data is Bls_multisig_counterBlsMultisigConfigType {
   if (!data || typeof data !== "object") return false;
   const obj = data as Record<string, unknown>;
 
@@ -1003,9 +1011,9 @@ export function isMulti_ibs_counterMultiIBSConfigType(
   return true;
 }
 
-export function isMulti_ibs_counterMultiIBSCounterType(
+export function isBls_multisig_counterBlsMultisigCounterType(
   data: unknown,
-): data is Multi_ibs_counterMultiIBSCounterType {
+): data is Bls_multisig_counterBlsMultisigCounterType {
   if (!data || typeof data !== "object") return false;
   const obj = data as Record<string, unknown>;
 
@@ -1024,9 +1032,9 @@ export function isMulti_ibs_counterMultiIBSCounterType(
   return true;
 }
 
-export function isMulti_ibs_counterMultiIBSProofType(
+export function isBls_multisig_counterBlsMultisigProofType(
   data: unknown,
-): data is Multi_ibs_counterMultiIBSProofType {
+): data is Bls_multisig_counterBlsMultisigProofType {
   if (!data || typeof data !== "object") return false;
   const obj = data as Record<string, unknown>;
 
@@ -1045,9 +1053,9 @@ export function isMulti_ibs_counterMultiIBSProofType(
   return true;
 }
 
-export function isMulti_ibs_counterPackageMarkerType(
+export function isBls_multisig_counterPackageMarkerType(
   data: unknown,
-): data is Multi_ibs_counterPackageMarkerType {
+): data is Bls_multisig_counterPackageMarkerType {
   if (!data || typeof data !== "object") return false;
   const obj = data as Record<string, unknown>;
 
@@ -1109,34 +1117,34 @@ export function isShared_counterSharedCounterType(
 }
 
 // Generated parse functions
-export function parseMulti_ibs_counterAggregatedPublicKey(
+export function parseBls_multisig_counterBlsAggregatedPk(
   json: unknown,
-): Multi_ibs_counterAggregatedPublicKeyType | null {
-  return isMulti_ibs_counterAggregatedPublicKeyType(json) ? json : null;
+): Bls_multisig_counterBlsAggregatedPkType | null {
+  return isBls_multisig_counterBlsAggregatedPkType(json) ? json : null;
 }
 
-export function parseMulti_ibs_counterMultiIBSConfig(
+export function parseBls_multisig_counterBlsMultisigConfig(
   json: unknown,
-): Multi_ibs_counterMultiIBSConfigType | null {
-  return isMulti_ibs_counterMultiIBSConfigType(json) ? json : null;
+): Bls_multisig_counterBlsMultisigConfigType | null {
+  return isBls_multisig_counterBlsMultisigConfigType(json) ? json : null;
 }
 
-export function parseMulti_ibs_counterMultiIBSCounter(
+export function parseBls_multisig_counterBlsMultisigCounter(
   json: unknown,
-): Multi_ibs_counterMultiIBSCounterType | null {
-  return isMulti_ibs_counterMultiIBSCounterType(json) ? json : null;
+): Bls_multisig_counterBlsMultisigCounterType | null {
+  return isBls_multisig_counterBlsMultisigCounterType(json) ? json : null;
 }
 
-export function parseMulti_ibs_counterMultiIBSProof(
+export function parseBls_multisig_counterBlsMultisigProof(
   json: unknown,
-): Multi_ibs_counterMultiIBSProofType | null {
-  return isMulti_ibs_counterMultiIBSProofType(json) ? json : null;
+): Bls_multisig_counterBlsMultisigProofType | null {
+  return isBls_multisig_counterBlsMultisigProofType(json) ? json : null;
 }
 
-export function parseMulti_ibs_counterPackageMarker(
+export function parseBls_multisig_counterPackageMarker(
   json: unknown,
-): Multi_ibs_counterPackageMarkerType | null {
-  return isMulti_ibs_counterPackageMarkerType(json) ? json : null;
+): Bls_multisig_counterPackageMarkerType | null {
+  return isBls_multisig_counterPackageMarkerType(json) ? json : null;
 }
 
 export function parseOwned_counterOwnedCounter(
@@ -1180,7 +1188,7 @@ if (!packageId) {
 export const counterPackage = {
   abi: counterAbi,
   packageId,
-  multi_ibs_counter: {
+  bls_multisig_counter: {
     aggregate_seal_shard_pubkey(
       tx: Transaction,
       options: {
@@ -1188,7 +1196,7 @@ export const counterPackage = {
       },
     ): TransactionResult {
       return tx.moveCall({
-        target: `${packageId}::multi_ibs_counter::aggregate_seal_shard_pubkey`,
+        target: `${packageId}::bls_multisig_counter::aggregate_seal_shard_pubkey`,
         arguments: [...options.arguments],
       });
     },
@@ -1199,24 +1207,24 @@ export const counterPackage = {
       },
     ): TransactionResult {
       return tx.moveCall({
-        target: `${packageId}::multi_ibs_counter::aggregated_key_counter_id`,
+        target: `${packageId}::bls_multisig_counter::aggregated_key_counter_id`,
         arguments: [...options.arguments],
       });
     },
-    destroy_aggregated_public_key(
+    destroy_bls_aggregated_pk(
       tx: Transaction,
       options: {
         arguments: readonly [TransactionArgument];
       },
     ): TransactionResult {
       return tx.moveCall({
-        target: `${packageId}::multi_ibs_counter::destroy_aggregated_public_key`,
+        target: `${packageId}::bls_multisig_counter::destroy_bls_aggregated_pk`,
         arguments: [...options.arguments],
       });
     },
     get_package_id_bytes(tx: Transaction): TransactionResult {
       return tx.moveCall({
-        target: `${packageId}::multi_ibs_counter::get_package_id_bytes`,
+        target: `${packageId}::bls_multisig_counter::get_package_id_bytes`,
         arguments: [],
       });
     },
@@ -1227,7 +1235,7 @@ export const counterPackage = {
       },
     ): TransactionResult {
       return tx.moveCall({
-        target: `${packageId}::multi_ibs_counter::included_seal_shard_count`,
+        target: `${packageId}::bls_multisig_counter::included_seal_shard_count`,
         arguments: [...options.arguments],
       });
     },
@@ -1238,7 +1246,7 @@ export const counterPackage = {
       },
     ): TransactionResult {
       return tx.moveCall({
-        target: `${packageId}::multi_ibs_counter::included_seal_shard_ids`,
+        target: `${packageId}::bls_multisig_counter::included_seal_shard_ids`,
         arguments: [...options.arguments],
       });
     },
@@ -1249,18 +1257,18 @@ export const counterPackage = {
       },
     ): TransactionResult {
       return tx.moveCall({
-        target: `${packageId}::multi_ibs_counter::increment`,
+        target: `${packageId}::bls_multisig_counter::increment`,
         arguments: [...options.arguments],
       });
     },
-    new_aggregated_public_key(
+    new_bls_aggregated_pk(
       tx: Transaction,
       options: {
         arguments: readonly [TransactionArgument];
       },
     ): TransactionResult {
       return tx.moveCall({
-        target: `${packageId}::multi_ibs_counter::new_aggregated_public_key`,
+        target: `${packageId}::bls_multisig_counter::new_bls_aggregated_pk`,
         arguments: [...options.arguments],
       });
     },
@@ -1271,7 +1279,7 @@ export const counterPackage = {
       },
     ): TransactionResult {
       return tx.moveCall({
-        target: `${packageId}::multi_ibs_counter::public_key_g2`,
+        target: `${packageId}::bls_multisig_counter::public_key_g2`,
         arguments: [...options.arguments],
       });
     },
@@ -1282,7 +1290,7 @@ export const counterPackage = {
       },
     ): TransactionResult {
       return tx.moveCall({
-        target: `${packageId}::multi_ibs_counter::seal_approve`,
+        target: `${packageId}::bls_multisig_counter::seal_approve`,
         arguments: [...options.arguments],
       });
     },
@@ -1293,7 +1301,7 @@ export const counterPackage = {
       },
     ): TransactionResult {
       return tx.moveCall({
-        target: `${packageId}::multi_ibs_counter::seal_shard_count`,
+        target: `${packageId}::bls_multisig_counter::seal_shard_count`,
         arguments: [...options.arguments],
       });
     },
@@ -1304,7 +1312,7 @@ export const counterPackage = {
       },
     ): TransactionResult {
       return tx.moveCall({
-        target: `${packageId}::multi_ibs_counter::seal_shard_table`,
+        target: `${packageId}::bls_multisig_counter::seal_shard_table`,
         arguments: [...options.arguments],
       });
     },
@@ -1315,7 +1323,7 @@ export const counterPackage = {
       },
     ): TransactionResult {
       return tx.moveCall({
-        target: `${packageId}::multi_ibs_counter::share`,
+        target: `${packageId}::bls_multisig_counter::share`,
         arguments: [...options.arguments],
       });
     },
@@ -1326,7 +1334,7 @@ export const counterPackage = {
       },
     ): TransactionResult {
       return tx.moveCall({
-        target: `${packageId}::multi_ibs_counter::threshold`,
+        target: `${packageId}::bls_multisig_counter::threshold`,
         arguments: [...options.arguments],
       });
     },
@@ -1337,7 +1345,7 @@ export const counterPackage = {
       },
     ): TransactionResult {
       return tx.moveCall({
-        target: `${packageId}::multi_ibs_counter::value`,
+        target: `${packageId}::bls_multisig_counter::value`,
         arguments: [...options.arguments],
       });
     },
@@ -1353,7 +1361,7 @@ export const counterPackage = {
       },
     ): TransactionResult {
       return tx.moveCall({
-        target: `${packageId}::multi_ibs_counter::verify_and_create_proof`,
+        target: `${packageId}::bls_multisig_counter::verify_and_create_proof`,
         arguments: [...options.arguments],
       });
     },
