+++ sui/src/abi/sui.abi.ts
@@ -3,7 +3,7 @@
  * 自動生成されたファイル - 手動で編集しないでください
  *
  * Generated by: scripts/suigenv0.ts
- * Generated at: 2025-09-27T03:52:05.781Z
+ * Generated at: 2025-09-27T08:26:57.886Z
  */
 
 import {
